Fixed CI docker build

See merge request dsantini/open-etymology-map!3

+++ web/db-init.php
@@ -19,6 +19,7 @@ $use_osm2pgsql = FALSE;
 $convert_to_geojson = FALSE;
 $convert_to_pg = FALSE;
 $convert_to_txt = FALSE;
+$keep_temp_tables = in_array("--keep-temp-tables", $argv);
 
 exec("which osmium", $output, $retval);
 if ($retval !== 0) {
@@ -52,7 +53,10 @@ $workDir = dirname($sourceFile);
 $sourceFilename = basename($sourceFile);
 echo "Working on file $sourceFilename in directory $workDir" . PHP_EOL;
 
-if (empty($argv[2])) {
+if ($keep_temp_tables)
+    echo 'Keeping temporary tables' . PHP_EOL;
+
+if (empty($argv[2]) || $argv[2] == "default") {
     echo 'Using osmium-export to pg (default)' . PHP_EOL;
     $use_osmium_export = TRUE;
     $convert_to_pg = TRUE;
@@ -98,34 +102,39 @@ function execAndCheck(string $command): array
     return $exOutput;
 }
 
-/**
- * @link https://docs.osmcode.org/osmium/latest/osmium-tags-filter.html
- */
+echo 'Started at ' . date('c') . PHP_EOL;
 
+$filteredTmpFile = sys_get_temp_dir() . "/filtered_with_flags_$sourceFilename";
 $filteredFile = "$workDir/filtered_$sourceFilename";
 if (is_file($filteredFile)) {
     echo '========================= Data already filtered =========================' . PHP_EOL;
 } else {
     echo '========================= Filtering OSM data... =========================' . PHP_EOL;
-    execAndCheck("osmium tags-filter --verbose --remove-tags --overwrite -o '$filteredFile' '$sourceFile' 'name:etymology:wikidata,subject:wikidata,wikidata'");
+    /**
+     * @link https://docs.osmcode.org/osmium/latest/osmium-tags-filter.html
+     */
+    execAndCheck("osmium tags-filter --verbose --remove-tags --overwrite -o '$filteredTmpFile' '$sourceFile' 'wikidata,subject:wikidata,name:etymology:wikidata'");
+    execAndCheck("osmium tags-filter --verbose --invert-match --overwrite -o '$filteredFile' '$filteredTmpFile' 'man_made=flagpole'");
     echo '========================= Filtered OSM data =========================' . PHP_EOL;
 }
 
-/**
- * @link https://docs.osmcode.org/osmium/latest/osmium-export.html
- */
 
 if (!is_file("osmium.json")) {
     echo "ERROR: missing osmium.json" . PHP_EOL;
     exit(1);
 }
 
+$cacheFile = tempnam(sys_get_temp_dir(), 'osmium');
+
 $txtFile = "$workDir/filtered_$sourceFilename.txt";
 if (is_file($txtFile)) {
     echo '========================= Data already exported to text =========================' . PHP_EOL;
 } elseif ($convert_to_txt) {
     echo '========================= Exporting OSM data to text... =========================' . PHP_EOL;
-    execAndCheck("osmium export --verbose --overwrite -o '$txtFile' -f 'txt' --config='osmium.json' --add-unique-id='counter' --index-type=dense_file_array,/tmp/osmium-nodes.cache '$filteredFile'");
+    /**
+     * @link https://docs.osmcode.org/osmium/latest/osmium-export.html
+     */
+    execAndCheck("osmium export --verbose --overwrite -o '$txtFile' -f 'txt' --config='osmium.json' --add-unique-id='counter' --index-type=sparse_file_array,$cacheFile '$filteredFile'");
     echo '========================= Exported OSM data to text =========================' . PHP_EOL;
 }
 
@@ -134,7 +143,10 @@ if (is_file($geojsonFile)) {
     echo '========================= Data already exported to geojson =========================' . PHP_EOL;
 } elseif ($convert_to_geojson) {
     echo '========================= Exporting OSM data to geojson... =========================' . PHP_EOL;
-    execAndCheck("osmium export --verbose --overwrite -o '$geojsonFile' -f 'geojson' --config='osmium.json' --add-unique-id='counter' --index-type=dense_file_array,/tmp/osmium-nodes.cache '$filteredFile'");
+    /**
+     * @link https://docs.osmcode.org/osmium/latest/osmium-export.html
+     */
+    execAndCheck("osmium export --verbose --overwrite -o '$geojsonFile' -f 'geojson' --config='osmium.json' --add-unique-id='counter' --index-type=sparse_file_array,$cacheFile '$filteredFile'");
     echo '========================= Exported OSM data to geojson =========================' . PHP_EOL;
 }
 
@@ -143,7 +155,10 @@ if (is_file($pgFile)) {
     echo '========================= Data already exported to PostGIS tsv =========================' . PHP_EOL;
 } elseif ($convert_to_pg) {
     echo '========================= Exporting OSM data to PostGIS tsv... =========================' . PHP_EOL;
-    execAndCheck("osmium export --verbose --overwrite -o '$pgFile' -f 'pg' --config='osmium.json' --add-unique-id='counter' --index-type=dense_file_array,/tmp/osmium-nodes.cache '$filteredFile'");
+    /**
+     * @link https://docs.osmcode.org/osmium/latest/osmium-export.html
+     */
+    execAndCheck("osmium export --verbose --overwrite -o '$pgFile' -f 'pg' --config='osmium.json' --add-unique-id='counter' --index-type=sparse_file_array,$cacheFile '$filteredFile'");
     echo '========================= Exported OSM data to PostGIS tsv =========================' . PHP_EOL;
 }
 
@@ -237,7 +252,9 @@ if ($use_db) {
                     --ew_id BIGSERIAL NOT NULL PRIMARY KEY,
                     ew_el_id BIGINT NOT NULL,
                     ew_wikidata_cod VARCHAR(15) NOT NULL CHECK (ew_wikidata_cod  ~* '^Q\d+$'),
-                    ew_etymology BOOLEAN NOT NULL
+                    ew_from_name_etymology BOOLEAN,
+                    ew_from_subject BOOLEAN,
+                    ew_from_wikidata BOOLEAN
                 )"
             );
             $dbh->exec(
@@ -268,6 +285,9 @@ if ($use_db) {
                     --et_el_id BIGINT NOT NULL REFERENCES oem.element(el_id), -- element is populated only at the end
                     et_el_id BIGINT NOT NULL,
                     et_wd_id INT NOT NULL REFERENCES oem.wikidata(wd_id),
+                    et_from_name_etymology BOOLEAN,
+                    et_from_subject BOOLEAN,
+                    et_from_wikidata BOOLEAN,
                     CONSTRAINT etymology_pkey PRIMARY KEY (et_el_id, et_wd_id)
                 )"
             );
@@ -361,6 +381,8 @@ if ($use_db) {
                 );
                 echo "========================= Converted $n_element elements =========================" . PHP_EOL;
             }
+
+            echo 'Loading complete at ' . date('c') . PHP_EOL;
         }
 
         if ($dbh->query("SELECT NOT EXISTS (SELECT FROM information_schema.tables WHERE table_schema='oem' AND table_name='element_wikidata_cods')")->fetchColumn()) {
@@ -371,16 +393,16 @@ if ($use_db) {
             } else {
                 echo '========================= Converting wikidata codes... =========================' . PHP_EOL;
                 $n_wikidata_cods = $dbh->exec(
-                    "INSERT INTO oem.element_wikidata_cods (ew_el_id, ew_wikidata_cod, ew_etymology)
-                    SELECT osm_id, UPPER(TRIM(wikidata_cod)), FALSE
+                    "INSERT INTO oem.element_wikidata_cods (ew_el_id, ew_wikidata_cod, ew_from_name_etymology, ew_from_subject, ew_from_wikidata)
+                    SELECT osm_id, UPPER(TRIM(wikidata_cod)), FALSE, FALSE, TRUE
                     FROM oem.osmdata, LATERAL REGEXP_SPLIT_TO_TABLE(osm_tags->>'wikidata',';') AS splitted(wikidata_cod)
                     WHERE TRIM(wikidata_cod) ~* '^Q\d+$'
                     UNION
-                    SELECT osm_id, UPPER(TRIM(subject_wikidata_cod)), TRUE
+                    SELECT osm_id, UPPER(TRIM(subject_wikidata_cod)), FALSE, TRUE, FALSE
                     FROM oem.osmdata, LATERAL REGEXP_SPLIT_TO_TABLE(osm_tags->>'subject:wikidata',';') AS splitted(subject_wikidata_cod)
                     WHERE TRIM(subject_wikidata_cod) ~* '^Q\d+$'
                     UNION
-                    SELECT osm_id, UPPER(TRIM(name_etymology_wikidata_cod)), TRUE
+                    SELECT osm_id, UPPER(TRIM(name_etymology_wikidata_cod)), TRUE, FALSE, FALSE
                     FROM oem.osmdata, LATERAL REGEXP_SPLIT_TO_TABLE(osm_tags->>'name:etymology:wikidata',';') AS splitted(name_etymology_wikidata_cod)
                     WHERE TRIM(name_etymology_wikidata_cod) ~* '^Q\d+$'"
                 );
@@ -392,27 +414,35 @@ if ($use_db) {
             } else {
                 $wikidataEndpointURL = (string)$conf->get("wikidata-endpoint");
 
-                $wikidataNamedAfterRQFile = "$workDir/wikidata_named_after.tmp.rq";
-                $wikidataNamedAfterJSONFile = "$workDir/wikidata_named_after.tmp.json";
+                $wikidataNamedAfterRQFile = sys_get_temp_dir() . "/wikidata_named_after.tmp.rq";
+                $wikidataNamedAfterJSONFile = sys_get_temp_dir() . "/wikidata_named_after.tmp.json";
 
                 $n_todo_named_after = $dbh->query(
-                    "SELECT COUNT(DISTINCT ew_wikidata_cod) FROM oem.element_wikidata_cods WHERE NOT ew_etymology"
+                    "SELECT COUNT(DISTINCT ew_wikidata_cod) FROM oem.element_wikidata_cods WHERE ew_from_wikidata"
                 )->fetchColumn();
                 echo "========================= Counted $n_todo_named_after Wikidata codes to check =========================" . PHP_EOL;
 
-                $pageSize = 10000;
+                $pageSize = 15000;
                 for ($offset = 0; $offset < $n_todo_named_after; $offset += $pageSize) {
                     echo "========================= Downloading Wikidata named-after data (starting from $offset)... =========================" . PHP_EOL;
                     $wikidataCodsToFetch = $dbh->query(
                         "SELECT STRING_AGG('wd:'||ew_wikidata_cod, ' ') FROM (
-                            SELECT DISTINCT ew_wikidata_cod FROM oem.element_wikidata_cods WHERE NOT ew_etymology ORDER BY ew_wikidata_cod LIMIT $pageSize OFFSET $offset
+                            SELECT DISTINCT ew_wikidata_cod FROM oem.element_wikidata_cods WHERE ew_from_wikidata ORDER BY ew_wikidata_cod LIMIT $pageSize OFFSET $offset
                         ) AS x"
                     )->fetchColumn();
                     $namedAfterQuery =
-                        "SELECT ?element ?namedAfter
+                        "SELECT DISTINCT ?element ?namedAfter
                         WHERE {
                             VALUES ?element { $wikidataCodsToFetch }.
-                            ?element wdt:P138 ?namedAfter.
+                            {
+                                ?element wdt:P138 ?namedAfter. # named after - https://www.wikidata.org/wiki/Property:P138
+                            } UNION {
+                                ?element owl:sameAs/wdt:P138 ?namedAfter.
+                            } UNION {
+                                ?element wdt:P825 ?namedAfter. # dedicated to - https://www.wikidata.org/wiki/Property:P825
+                            } UNION {
+                                ?element owl:sameAs/wdt:P825 ?namedAfter.
+                            }
                         }";
                     file_put_contents($wikidataNamedAfterRQFile, $namedAfterQuery);
                     $jsonResult = (new JSONWikidataQuery($namedAfterQuery, $wikidataEndpointURL))->sendAndGetJSONResult()->getJSON();
@@ -457,7 +487,7 @@ if ($use_db) {
                     SELECT DISTINCT ew_wikidata_cod
                     FROM oem.element_wikidata_cods
                     LEFT JOIN oem.wikidata ON wd_wikidata_cod = ew_wikidata_cod
-                    WHERE ew_etymology
+                    WHERE (ew_from_name_etymology OR ew_from_subject)
                     AND wd_id IS NULL"
                 );
                 echo "========================= Loaded $n_wd Wikidata entities =========================" . PHP_EOL;
@@ -470,22 +500,29 @@ if ($use_db) {
             } else {
                 echo '========================= Converting etymologies... =========================' . PHP_EOL;
                 $n_ety = $dbh->exec(
-                    "INSERT INTO oem.etymology (et_el_id, et_wd_id)
-                    SELECT DISTINCT ew_el_id, wd_id
-                    FROM oem.element_wikidata_cods
-                    JOIN oem.wikidata ON ew_wikidata_cod = wd_wikidata_cod
-                    WHERE ew_etymology
-                    UNION
-                    SELECT DISTINCT ew.ew_el_id, nawd.wd_id
-                    FROM oem.element_wikidata_cods AS ew
-                    JOIN oem.wikidata AS wd ON ew.ew_wikidata_cod = wd.wd_wikidata_cod
-                    JOIN oem.wikidata_named_after AS wna ON wd.wd_id = wna.wna_wd_id
-                    JOIN oem.wikidata AS nawd ON wna.wna_named_after_wd_id = nawd.wd_id
-                    WHERE NOT ew.ew_etymology"
+                    "INSERT INTO oem.etymology (et_el_id, et_wd_id, et_from_name_etymology, et_from_subject, et_from_wikidata)
+                    SELECT ew_el_id, wd_id, BOOL_OR(ew_from_name_etymology), BOOL_OR(ew_from_subject), BOOL_OR(ew_from_wikidata)
+                    FROM (
+                        SELECT DISTINCT ew_el_id, wd_id, ew_from_name_etymology, ew_from_subject, ew_from_wikidata
+                        FROM oem.element_wikidata_cods
+                        JOIN oem.wikidata ON ew_wikidata_cod = wd_wikidata_cod
+                        WHERE ew_from_name_etymology OR ew_from_subject
+                        UNION
+                        SELECT DISTINCT ew.ew_el_id, nawd.wd_id, ew_from_name_etymology, ew_from_subject, ew_from_wikidata
+                        FROM oem.element_wikidata_cods AS ew
+                        JOIN oem.wikidata AS wd ON ew.ew_wikidata_cod = wd.wd_wikidata_cod
+                        JOIN oem.wikidata_named_after AS wna ON wd.wd_id = wna.wna_wd_id
+                        JOIN oem.wikidata AS nawd ON wna.wna_named_after_wd_id = nawd.wd_id
+                        WHERE ew_from_wikidata
+                    ) AS x
+                    GROUP BY ew_el_id, wd_id"
                 );
-                $dbh->exec("DROP TABLE oem.element_wikidata_cods");
+                if (!$keep_temp_tables)
+                    $dbh->exec("DROP TABLE oem.element_wikidata_cods");
                 echo "========================= Converted $n_ety etymologies =========================" . PHP_EOL;
             }
+
+            echo 'Conversion complete at ' . date('c') . PHP_EOL;
         }
 
         if ($dbh->query("SELECT NOT EXISTS (SELECT FROM information_schema.tables WHERE table_schema='oem' AND table_name='osmdata')")->fetchColumn()) {
@@ -498,7 +535,8 @@ if ($use_db) {
             );
             $n_remaining = $n_tot - $n_cleaned;
             $dbh->exec('ALTER TABLE oem.osmdata RENAME TO element');*/
-            $n_remaining = $dbh->exec( // About 90% of elements is deleted, it's faster to copy them; https://stackoverflow.com/a/7088514/2347196
+            // Almost 90% of elements would be deleted, it's faster to copy them in another table; https://stackoverflow.com/a/7088514/2347196
+            $n_remaining = $dbh->exec(
                 "INSERT INTO oem.element (el_id, el_geometry, el_osm_type, el_osm_id, el_name)
                 SELECT osm_id, osm_geometry, osm_osm_type, osm_osm_id, osm_tags->>'name'
                 FROM oem.osmdata
@@ -506,7 +544,9 @@ if ($use_db) {
             );
             $n_cleaned = $n_tot - $n_remaining;
             echo "========================= Cleaned up $n_cleaned elements without etymology ($n_remaining remaining) =========================" . PHP_EOL;
-            $dbh->exec('DROP TABLE oem.osmdata');
+            file_put_contents('LAST_UPDATE', date('Y-m-d'));
+            if (!$keep_temp_tables)
+                $dbh->exec('DROP TABLE oem.osmdata');
         }
 
         echo '========================= Generating global map... =========================' . PHP_EOL;
@@ -534,4 +574,6 @@ if ($use_db) {
     } catch (Exception $e) {
         echo "ERROR:" . PHP_EOL . $e->getMessage() . PHP_EOL;
     }
+
+    echo 'Finished at ' . date('c') . PHP_EOL;
 }
